Move start-account-or-login to lib directory.
