+++ frontend/lib/norent/letter-builder/routes.tsx
@@ -5,7 +5,7 @@ import {
 import { NorentRoutes } from "../route-info";
 import { NorentLandlordNameAndContactTypes } from "./landlord-name-and-contact-types";
 import { NorentLetterPreviewPage } from "./letter-preview";
-import { createStartAccountOrLoginSteps } from "../start-account-or-login/routes";
+import { createStartAccountOrLoginSteps } from "../../start-account-or-login/routes";
 import { AllSessionInfo } from "../../queries/AllSessionInfo";
 import { NorentLetterBuilderRouteInfo } from "./route-info";
 import { NorentLbWelcome } from "./welcome";
