+++ frontend/lib/start-account-or-login/route-info.ts
@@ -3,7 +3,7 @@ export type StartAccountOrLoginRouteInfo = ReturnType<
 >;
 
 /**
- * This function maps URL paths to our routes within the NoRent Account Creation/Login flow.
+ * This function maps URL paths to our routes within Account Creation/Login flow.
  * To find the actual definition of these routes, check out
  * the `routes.tsx` file in the same directory as this file.
  */
